Fix saving preferences of type enum with empty value
when a preference with empty value is saved, this row was just
deleted from the db. This is not true for enums, because an empty string
is not always the default value. Therefore we only delete enum preferences
when their value matches the one of the enum, configured as default.

This can be checked, by setting the domain preference outbound_pai_user
to None, which resulted in wrong behaviour before.

Mantis: 3769