Show Timeset and Destination Set Names
and provide a Modal showing the Detail of each.

Except the case when thedestination_set.name is called "quickset_*" and
it only has one entry.
In that case show the particular destinations of this destination set,
in order to not confuse a user who just configured a trivial cf via the simple view.

under Subscriber Preferences -> Call Forwards

Mantis: 3803, 3801