MT#4289 contracts: filter by product and separate terminate
Only show reseller and peering contracts in the
Contracts controller. Also remove the generic create
in favour of the peering/create and reseller/create.

Customer controller now has its own terminate routine to be really
independent from the Contract controller.

Now we can expose the Contract controller in the menu.