MT#3855 Fix empty contract tab on Reseller details
When the reseller has a wrong contract assigned to
him, this contract is not shown because it has already
been filtered out in the Contract controller. In favor
of usability now show it anyway.

Also only present correct contracts on edit/create
resellers in the first place.

See (Mantis:3855#c25797)