MT#4955 Introduce customer default sound sets.
If a new contract sound set is marked as contract_default, use this
for every existing subscriber in contract_sound_set prefs if no
other set is specified yet.
If a new subscriber is created and such a sound set exists, also
use this one.

This prevents having to go over each single subscriber to assign
the contract_sound_set preference.