MT#5661 Update prepaid pref on bilprof change
If prepaid flag is changed in billing profile, all subscribers
of all contracts currently having this profile are updated.

If customer changes from prepaid profile to postpaid and vice versa,
all subscribers of this contract are updated.

If subscriber is created either in subscriber or customer view,
the prepaid pref is set if billing profile of its customer is
prepaid.