Update patch set 1
Patch Set 1: Code-Review+1

ACK for me. IMHO it is not final version for mr3.6 just a step to move forward. IMHO we need to add 'realm' to nginx config and pass it to ngcp-panel through fastcgi_parameter. So, ngcp-panel will know who is comming admin/reseller/subscriber...

Patch-set: 1
Label: Code-Review=+1