Update patch set 3
Patch Set 3:

How does this work in production?

If you have a model with linerange A and linerange B, and have field devices using only linerange A, and you do an update on the model to remove linerange B, then the update would have worked before already by just removing linerange B, right?

Now if field devices use both linerange A and linerange B, and you delete linerange B via the API, then the current code would silently NOT delete this linerange? That's probably not what is expected.

If you update the model, and you explicitely REMOVE a linerange, then all field devices should have their subscribers removed from the config which is using this linerange.

However, if you just UPDATE a linerange, then the subscribers should still be kept. Lineranges are (or should be) uniquely identified by their name, so if you modify a model and keep the name of the line ranges as it was (but only add more lines), then the subscribers of the field devices should be kept. If you remove lines from a line range and you have subscribers on this lines, the subscribers should be removed.

The only thing we REALLY need to make sure is that on a normal update of a model which doesn't really change the lineranges, the subscribers should be kept on the field devices.

I think the best thing for that cases to validate is to create API tests in t/api-deviceprov.t for example and create device models there, create field devices with subscribers using those line ranges, then start to update/delete line ranges.

Patch-set: 3
Reviewer: Andreas Granig <1000007@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0