Update patch set 3
Patch Set 3:

My thought was that such valued action as subscriber line deletion should be done only explicitly, and not by updating the Model. It is very easy to create incorrect request and loose subscriber lines by accident. As a variant, it can be implemented as error: if client tries to update Model and it will lead to loose of some subscriber lines - we can show a error, to don't be silent.

Patch-set: 3
Reviewer: Irina Peshinskaya <1000009@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0