Update patch set 1
Patch Set 1:

Hello )
I added you to this, the same as to previous review from backporting policy point of view. I couldn't push this into 3.6 directly, this is the only reason to review these changes. From Perl point of view it really doesn't worth review, at least because this fix is presented in the master from August, but together with other changes.

Patch-set: 1