Update patch set 2
Patch Set 2:

> ltgtm, a note about not_nullable: imo we shouldn't set it usually,
 > assuming "NOT NULL" in mysql intends to say neither null nor empty.

full ack - that "not_nullable" was already in place there, but again i didn't dare to replace it with required=>1.

"assuming "NOT NULL" in mysql intends to say neither null nor empty." that is exactly what i think as well - and that's the yet unresolved thing with the mysql server: it auto-defaults to empty string in my local db, irina's local db, my vagrant ce-trunk VM but NOT in the ce-trunk VM jenkins uses to run the tests (which is the correct behaviour after all). would like to reproduce/clarify this though..

Patch-set: 2
Label: Code-Review=0