Update patch set 1
Patch Set 1:

The fix was initially requested by UPC-HPBX where for calls inside the same pbx groups (intra) the caller should not be anonymised.

A cut from ticket 13311:

> So to be clear:
> 
> direction = "in", intra_customer = "false", "source_clir" = 1 -- other_cli = undef
> direction = "in", intra_customer = "true", "source_clir" = 1 -- other_cli = extension

This is perfect. 

I used the opportunity to do a quick test, and it now seems to be working correctly.

Patch-set: 1
Label: Code-Review=0