MT#13903 disable applying profile package
This is a patcht introducing a variable to disable
setting profile_package_id's in in the contracts
table. The code should therefore behave as before,
as long rate-o-mat and libswrate are ready.

Not once set to 1 and contract balances are written
to a database once, disabling this constant will
mess up contract_balances!

Change-Id: I080460395627202f3f5a386022de643577f4f3c2