Update patch set 7
Patch Set 7:

lgtm.. in general, i think thats the way to do it - first retrieve graph to clone and subsequently insert records again. btw. those "duplicate" buttons are likely to be helpful in some other situations as well. i would not opt for a generalized approach for cloning either.

Patch-set: 7