Update patch set 2
Patch Set 2:

Review raised because I was (and am now)in doubt if it is really necessary to backport it to 4.0.1 - customer closed his own ticket, and seems like the fix in the next release would be also OK for him, it is just documentation.
So it is a question of the backporting Documentation enhancement  - to do or not to do?

Patch-set: 2
Reviewer: Irina Peshinskaya <1000009@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0