MT#14317 passing now() to interval catchup and resize
formerly, the contract_balance catchup (and resize) procedures
used contracts.modified_timestamp, whose value was intended to
be set explicitly upon update operations.

it turns out that mysql ON UPDATE current_timestamp columns
effectively ignore the explicit timestamp value, unless it
exactly matches the old timestamp value.

while this is not an issue for the logic in production usage,
this seems to be the reason why the api-balance_intervals.t test
fails occassionally with its time-warp cases.

Change-Id: I504dd98548b88c99079e071598806d884c4422eb