MT#14255 add cash_balance underrun logic
in short: whenever the contract_balances "cash_balance"
is about to decrease, it might drop below thresholds
defined by a contract's profile package. in that case,
subscribers are to be locked and/or underrun billing
profile sets are to be applied.

this commit also contains some minor fixes:

+when creating/editing profile packages, all billing
profiles of initial, topup, underrun sets must show
the same interval_free_cash and interval_free_time
values

+when editing packages using the panel UI, the
validation was not properly invoked (!)

+the frequency of topups has to be limited to 1
per second.

still missing is updating the "lock" subscriber prior
to editing or viewing it in the panel UI.

Change-Id: Ideb0fec2404ca10b3ac5d8735f53620d0c3a24ed