Update patch set 1
Patch Set 1:

This is an initial docker setup for usage with the api-rest tests (it's based on top of https://gerrit.mgm.sipwise.com/#/c/2667/).

It would be great if someone of you could look at the known problems I've mentioned in the commit message. If there's anything unclear or if I could be of any help please let me know.

Patch-set: 1
Label: Code-Review=0