Update patch set 1
Patch Set 1: Code-Review+2

as discussed here https://gerrit.mgm.sipwise.com/#/c/2679

release hotfix for mr3.8.3, as it is require for netcologne and we are updating netcologne next week to mr3.8.3

in the same time IMHO hotfix here is too dangerous (changes are to big and touches a lot of _different_ places). IMHO, it should be in mr3.8.4+ (cherry-picked to mr3.8 branch only).

+2 if you still would like to have hotifx here.

Patch-set: 1
Label: Code-Review=+2