Update patch set 1
Patch Set 1:

I think there is nothing to be sorry for.
I think that it is question of the support team and QA to figure out the process to deliver requested features to the customers, which don't plan to upgrade in the nearest future. 
I can't predict all points which should be considered by this process.
From my side I'll integrate in any branch which will be required by support team with respect to the QA requirements & rules.

Patch-set: 1