Update patch set 5
Patch Set 5: Code-Review+1

the basic problem is that for " OR " statements, mysql will reformulate that into UNION, which gets inacceptable slow under the circumstances andrew pointed out originally. the reason behind is that UNION removes dupes like DISTINCT does, which is the culprit since this raises the query time up to minutes, if a call list is large enough.

so a proper approach would be to somehow force UNION ALL instead of union, which does not involve the DISTINCT filter. with that approach we also accept to have duplicate rows in the in+out call list for self-calls. this would also break conventions to match the api's output, but it's likely that it will be the only really performant way to overcome that situation.

so as a first step, we (andrew, rene, ira) decided to really just change the default filters of the call list view, to prevent stuck panel UI screens and frequent long running queries when users just open the call list view initially.

Patch-set: 5
Label: Code-Review=+1