Update patch set 1
Patch Set 1:

That's correct. But do you actually use it when translating something? Or isn't it just as easy to search the specific string (the filename is still there).

The reason was to prevent the big regular commits by jenkins, which happen, when something is changed in a large files; and therefore a lot of line numbers get updated.

Anyway, it was just a suggestion and nothing critical. So if you think the line-numbers are useful, we'll keep them. So what do you say?

Patch-set: 1
Label: Code-Review=0