Update patch set 1
Patch Set 1:

This really looks like a heavy violation of REST to me. Also it doesn't align well with the rest of our API.

I understand that for backwards-compatibility you cannot just remove PUT (although for an new version, users should be able to deal with that).

So if someone else approves this, it's fine by me, I just wanted to say my concerns.

Patch-set: 1