Update patch set 12
Patch Set 12:

(5 comments)

since the prior in/out filtering was relying on


literal_sql => 'if(source_user_id = "'.$c->stash->{subscriber}->uuid.'", "outgoing", "incoming")' }


which is not accurate either, i think we can accept showing self-calls twice, as shown in the line comments. also the api callist filter should could be adopted to that.

what do you think? partitioning won't solve that, but with union_all we will always be on the safe side. we should finish this.

Patch-set: 12
Label: Code-Review=0