Update patch set 3
Patch Set 3:

if DEBUG is on (it is on now by default) there is logging of pretty much everything. If it stops afte "Loading fax preferences" and nothing else is shown then there is something wrong. Let me know please. By the way Alex tested it a few weeks ago just out of the box (freshly baked vagrant) and it worked without issues (granted with the call forwardings, etc. configured for subscribers).

Patch-set: 3