Update patch set 4
Patch Set 4:

Not sure what the customer wants by "receive fax" (currently we have only fax2mail). 
As for the "cover" pages support, yes hylafax does have it, although we did not use it.
I can implement "cover" pages support, some other "receive faxes" methods in the new ngcp-faxserver, no problem, but the question is. Is the customer going to pay us for all the new feature requests ?
Currently ngcp-faxserver supports the core set of features we had with hylafax + some more. Any other new features in ngcp-faxserver are going to be implemented through either scheduling them by Andreas or as paid feature requests.
If there was something that existed in hylafax and we do not have it in ngcp-faxserver and it makes sense to implement similar functionality - I will do that + of course bugfixing/polishing, etc. will be done as the normal dev routine.
Things like "oh, there was a port listening 4559/tcp and we could connect to it with a fax client" - we now have mail2fax feature which is way more comfortable for that than finding and installing a standalone client, dealing with firewalls, etc.

Patch-set: 4