Update patch set 4
Patch Set 4:

Irina, could you please prepare and send me a list of features we promised the "customer" would have. If there is something that was promised and missing in ngcp-faxserver, I will implement it asap.
Things like "receive fax" method, we did not have it in hylafax nor in the old faxserver, did we promise it to the customer ?
Cover pages support, we had it in hylafax, but did not use in the old faxserver, did we promise it to the customer ?

Patch-set: 4
Label: Code-Review=0