MT#15269 fix some api tests, add test selection
by choosing "stable" (about 1m30s) or "fast" (about 30s) as second
argument of testrunner, only tests are run, which are currently
succeeding on a plain trunk vm using docker. (times are on my machine)

I hope we can in this way re-enable them on nightlies (or even in gerrit)
so that devs have some better feedback on the impact of our changes.

Current status of tests (in trunk):

| Problems:
| 	api-callforwards.t
| 	api-balanceintervals.t
| 	api-faxes.t
| 	api-journals.t
|
| Fixed:
| 	api-all-links.t
|       api-contracts.t
|       api-customercontacts.t
|       api-customers.t
| 	api-rewriterulesets.t
| 	api-root.t
|
| Working:
| 	api-billingfees.t
| 	api-billingnetworks.t
| 	api-billingprofiles.t
| 	api-billingzones.t
| 	api-calllists.t
| 	api-calls.t
| 	api-cert-auth.t
| 	api-cfdestinationsets.t
| 	api-ncoslevels.t
| 	api-pbxdevicemodels.t
| 	api-pbxdevices.t
| 	api-peeringgroups.t
| 	api-peeringrules.t
| 	api-peeringservers.t
| 	api-preferences.t
| 	api-profilepackages.t
| 	api-resellers.t
| 	api-rewriterules.t
| 	api-soundsets.t
| 	api-subscriberregistrations.t
| 	api-subscribers.t
| 	api-systemcontacts.t
| 	api-threads.t
| 	api-topuplogs.t
| 	api-trustedsources.t
| 	api-valid-patch.t
| 	api-vouchers.t

Change-Id: I74e672d9dc6bcccfd3ace2fe233dbcf325c6fc4d