Update patch set 5
Patch Set 5:

I think I will just "require" NCP::Fax based on facserver feature enabled in the config.
Is it still correct to add NGCP::Fax as dependency in the panel installation config? I'm in doubts, because I didn't see conditions there in the debian/control, but in the same time ce will not have this feature, if I get this right, and so don't need NGCP::fax  package too.

Patch-set: 5