Update patch set 11
Patch Set 11:

> Just wanted to ask - did you find Collection inconvenient for this
 > case? Why? Or maybe you find it unreliable?

Yes, basically this is just faster for me to write. For me it's a bit difficult to understand the Collections framework, whereas here I am in complete control of what's happening.

Ideally imo, now that we have the API client library, we could consider using that. (NGCP::API::Client in ngcp-api-tools)

Patch-set: 11
Label: Code-Review=0