Update patch set 1
Patch Set 1: Code-Review+1

lgtm, good catch

for my understanding (it's not in the handbook): why can we set profile preferences and which ones can be set? are they supposed to be a kind of default? then wouldn't it make more sense, to set those, which can NOT be edited by a subscriber?

Patch-set: 1
Reviewer: Gerhard Jungwirth <1000005@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+1