Update patch set 1
Patch Set 1:

regarding actual fix: the logic basically implements an efficient set-substraction method using hashmaps, to identify which preferences to finally drop from db. and there was only a "typo" in the line that builds the initial set. meantime volodymyr confirmed it fixes the problem as expected.

regarding reseller_edit config option (my assumption from reading the code): since editing a subscriber profile is a "dangerous" operations (it will potentially delete a large number of subscriber preference values if the pref itsself is unchecked from the profile), the idea was to have a "reseller_edit" option to leave pref values untouched. if reseller_edit is enabled, resellers are only allowed to change the profile's name and the like.

Patch-set: 1