MT#19795 rtc api consider app id for sessions
a reseller can manage multiple apps. these will have different networks in the
future, enabling one to allow or disallow networks depending on that.
in that case, the subscriber must have the possibility to select, which app they
are going to use. if none is selected, the default app is used.

Change-Id: I191bf3911004d405086821a89b46e4fa59379a46