Update patch set 6
Patch Set 6:

Yeah debian/copyright would need to be updated. Checking briefly the upstream code, it seems there are several others Copyright holders, not just the one listed on the LICENSE file. I guess a «grep Copyright» would reveal them.

Another question is if embedding this inside the ngcp-panel is the best of the options, instead of packaging it independently, which of course has the possible problem that might be more difficult as that would include not just the editor directory there. Hmmm.

Patch-set: 6
Label: Code-Review=0