Update patch set 1
Patch Set 1:

> lgtm. But we need to keep it in mind, because sometimes customers
 > put csc and admin to different hosts

even better. actually, this is the only way to really separate the two interfaces. because due to the way, web browsers work, cookies will always be shared between different ports. so even with this change, one can log in as admin and then change to url to the csc manually. there is nothing we can do about that

Patch-set: 1