Update patch set 7
Patch Set 7:

> > Yeah debian/copyright would need to be updated. Checking briefly
 > the upstream code, it seems there are several others Copyright
 > holders, not just the one listed on the LICENSE file. I guess a
 > «grep Copyright» would reveal them.
 > 
 > 
 > I'm afraid, here I will need more help.
 > As I see ngcp-panel/debian/copyright now doesn't contain list of
 > the authors. So what is the course for my changes in this file? Do
 > I need to add list of the authors from the scg-edit?

The authors are not necessarily relevant, the important information here is who is the copyright holder. If we stay with embedding the code here, you'd append a new paragraph, such as:

Files:
 share/static/js/libs/svg-edit/*
Copyright:
 Copyright © 2005-2015 Some Name <some@example.org>
 Copyright © 2006-2014 Other Name <other@example.org>
License: MIT
 <Here just license part of the LICENSE file, with blank lines
 replaced with a dot, such as
 .
 this previous one.>

As mentioned before you can probably just extract the list of copyright holders with a simple «git grep Copyright» on that specific tree.

If going with an independent package, then the debian/copyright would contain the format paragraph and that single license paragraph instead of two.

Patch-set: 7
Label: Code-Review=0