Update patch set 9
Patch Set 9:

Ok, I've discussed this with Victor, and given that there are tons of other embedded code copies in here, I guess it does not make sense (for now) to be picky about this one. Also when I see this is already in the repo but probably an older version. So I guess squashing the other related commit (thar removes extensions and fixups the copyright  file) with this one would be ideal, and then can be merged.

In addition the above realization means the copyright file probably does not match reality and might need to be updated to cover all those embedded code copies. IMO the ideal solution mid/long-term would be to split all those into independent source packages, but that might imply quite some work. :/

Patch-set: 9
Label: Code-Review=0