Update patch set 2
Patch Set 2: Code-Review+2

thx!
but to me it seems, this is only one part of the whole solution. in particular, the the query in the sems config must be extended to implement the same selection. and the latter is the actual urgent part, reguired for the melita mig :-)

Patch-set: 2
Reviewer: Rene Krenn <1000013@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2