Update patch set 3
Patch Set 3:

I changed here logic a little for lnpnumbers update validation:

Now we allow the same code in the same item. As usually. Before we returned error on attempt to set code to the same value.

Is it OK? If no objections, I will commit it tomorrow. 

I think it should be merged in 4.5.1 too.

Patch-set: 3