Update patch set 1
Patch Set 1: Code-Review+2

LGTM... commit message can be better, while it is good enough for the first commit.

P.S. as I can see it is a bit out dated, while lets merge it and then finish the missing one. Anyway the task which upgrade the en.po in trunk is broken too.

Patch-set: 1
Label: Code-Review=+2