Update patch set 3
Patch Set 3:

I wouldn't store this information into a cookie. Especially, since this information is by no means tied to our sessions. Therefore, what one user sets could cause unexpected results for another user (or even when switching to the next controller). This could be problematic when e.g you set it to all, and switch to a page with a high number of rows.

Datatables already has a settings-persistence built in. We are already utilizing it, using localStorage. imo, this should be extended to your needs (implementing a longer persistence, or a different application scope of individual settings)

Patch-set: 3
Label: Code-Review=0