Update patch set 3
Patch Set 3:

localstore is html5 only to my knowledge, but i guess the only way to solve this: cookie size limit becomes a problem if you store these setting for our numerous datatables, at least that is what i went into once. so a pagesize (and other) values per datatable-name+ngcp usename are read initially when the page loads and written once the user changes the pagesize. this is also the usual approach in ExtJs

Patch-set: 3