Update patch set 4
Patch Set 4:

ok, so the formhandler shows an inflate/deflate mechanism to convert it to the "..avp.." notation using regexp... and the api formhandler inherits from panel ui form handler. but it does not explain why it works in api but not in panel.. finally there are lots more "avp" regexpes, what else is affected?

Patch-set: 4