Update patch set 1
Patch Set 1: Code-Review+2

> Customer has installed 4.5.2

it is not a reason for the hotfixes in general!
mr4.5.3+ has the fix, please schedule the upgrade.

> The fix is pretty urgent for him.

Ok for this time... while PLEASE minimize hotfixing when possible,
moreover if is is not explicitly requested.

Patch-set: 1
Reviewer: Alexander Lutay <1000001@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2