Update patch set 1
Patch Set 1:

Disclaimer: I'm not really sure that this is the proper fix since I'm not familiar with ngcp-panel's code at all and what's it supposed to do, nor do I know any details about the SVG format.

So please take care with my review, if the "\{|\}" is really the proper fix or if that was supposed to be read as "|" only or so. :) Feel free to update my review with a corrected patchset.

FYI: with this change applied ngcp-panel starts on Debian/stretch, that's something I consider great news. :)

Patch-set: 1