TT#12660 Also send latin1 as gsm encoding.
Except for greek characters, which are available in the
GSM 03.38 encoding but not in latin-1, all other latin-1
characters can be sent with coding 0 (7bit).

We neither support the € sign available in the GSM extended
encoding, but for unknown reasons the rest of the extended
char sets (which are part of ascii anways) do work.

So, no € and no greek chars inside coding=0.

Change-Id: Ia87e337772e126b6a0a95b53acba0a369b71e660