TT#12601 Move admin pass to bcrypt and drop cert
For the migration of the admin pwd, the logic is as follows:

1. If the admin has a bcrypt password already, use this
2. If not, perform auth via md5, then clear the md5 column
   and write a salted bcrypt hash instead.

For dropping the ssl client cert, we simply not store anymore
the client certificate in the DB. As a result, you cannot
download the certs (pem, p12) after creation anymore, so
we immediately download the two certs in a zip file after
creation.

A cost of 13 takes 500ms on an i7-5500U CPU @ 2.40GHz, which
seems to be a reasonable value.

Change-Id: I1ce21321c58d8c57d7ddce1541995f64821b0053