TT#13249 Implement API to create admin client cert
Admins can use this to generate a new ssl client certificate package
in ZIP format containing a PEM and a P12 cert, the same way you can
do it on the admin panel.

Admins can generate that for themselves in all cases, or for others
if they are master or superuser. If role is reseller, then masters
can only do this within their own reseller.

Change-Id: I32d5c1b5af5324d1c80b34bacecd7f2665cd91c7