TT#5420 Depend on the now split libngcp-rating-inew-perl package
These modules used to be incorrectly shipped in the ngcp-libinewrate0
package, which the panel didn't depend on, and were only present because
other parts of the system were pulling that in. In any case we should
not need to depend on specific SONAMEs for shared libraries, so the perl
bindings are now shipped on their own package.

Reintroduce reverted change.

Change-Id: Ia6416fdbf8337549a146ae15ce0dd3f77e720ff3