TT#5420 Depend on the now split libngcp-rating-inew-perl package
This used to be incorrectly shipped in the ngcp-libinewrate0 package,
which the panel didn't depend on, and was only present because other
parts of the system were pulling that in. In any case we should not
need to depend on specific SONAMEs for shared libraries so the perl
bindings are now shipped on their own package.

Change-Id: Ia4ac5258f4dc019fecd1f138d7c0e6f6390296d0