Update patch set 2
Patch Set 2:

sleep:

yes and no. this is anyway not yet ready for automated testing because of the stated remaining issues. but it's a step forward and at least helps me in automated testing. therefore I would also leave the sleeps in for the moment.

also, I don't have so much time to work on a proper solution immedeately.

there is however one more issue with geckodriver that should be mentioned which also would aid in this problem: it seems the setting of the wait-timeout has no effect. this seems to be another incompatibility between S:R:D and geckodriver/the w3c specs. I'll try a fix...

Patch-set: 2