Update patch set 1
Patch Set 1:

it could be made configurable, but oliver only wants to have it removed in 5.4 to my understanding. it will be permanent again in future. the reason i introduced the enable constant at all was to test the difference in speed for large callists. that constant now got useful in some way.

there are no testcases for it so long, as there is no api to register these supression regexpes for now (lets see if anyone will use the feature at all...). but all affected panel parts (subscriber widget, call lists) were tested thoroughly with all combinations.

Patch-set: 1
Label: Code-Review=0