TT#18115 Call start-stop-daemon correctly
We should pass the interpreter executable as --exec and the actual
script to run as --startas, as documented in the man page. Otherwise
start-stop-daemon will not be able to match the executable currently
running, and will consider that the daemon is *not* running. Use
--oknodo to handle already running daemons.

Use start-stop-daemon instead of stopping the daemon by hand.

Change-Id: I026fda70dfaeb259afe5c7905c612dce6affd46e