TT#21767 Use git version of Selenium::Remote::Driver
Selenium::Remote::Driver v1.20 is the latest upstream release and
is missing some patches/PRs that we need for selenium tests of
ngcp-panel. Packaging the git version properly in the Debian way
is hard because of the way upsteam handles releases with separate
branches (one of them even being out-of-date, sigh). So instead
let's use a tarball from their latest git version until we see a
v.1.21 release.

Symlinking the geckodriver binary to /usr/bin/geckodriver addresses:

| Unable to find the geckodriver binary in your $PATH. at /home/selenium/Selenium-Remote-Driver/lib/Selenium/CanStartBinary/FindBinary.pm line 66.

Change-Id: I8ddf3c76a008ce5b0aa67b62154a9614f90cb659