Update patch set 2
Patch Set 2: Code-Review-2

I am just trying to figure out if this is the reason of failing kamailio-config-tests, since we made some other changes in kamailio too. Gerrit added you to the review automatically, I'm really sorry for the noise. No, i do not plan to submit this. I put myself -2 as it is only for testing...

Patch-set: 2