Update patch set 1
Patch Set 1: Code-Review+2

this one is testcase related only:
the new fields are safe for put2get cases - there must not be a 422 when the id's are included in a put/patch.

but primary_number_id had been removed, which we definitely need to reintroduce as customers might complain.

so ok for this particular one i would say.

Patch-set: 1
Reviewer: Rene Krenn <1000013@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2