Update patch set 1
Patch Set 1:

i think if we take out the new fields, andreas' new tests might fail.

this one here really just popped up because ira is doing a deep compare of e164 hashes. those have a new field now so the test code fails - ever since deep compares should only consider the fields that are really about to be compared.

the other commit needs to be placed in any case, and it wont fix this one - originally i thought this is about a patch 422 error due to the new fields (but isnt)

Patch-set: 1