Update patch set 2
Patch Set 2:

We do NOT backport. we force users upgrade. In the first case we perform work which creates more work for us (further bugfixing on backported code).

It has been clarified a LOT of times. Sure customer's will ask us to do this while we are releasing new releases VERY often to be able to deliever every new feature ASAP.

Patch-set: 2