Update patch set 8
Patch Set 8:

looks way better now, thank you very much Irina!
please keep the review open, I am going to help you with the next steps of moving it into the ngcp-dev-tools repo, and the packaging part, as well as some of the code further improvements in terms of module/client parts, those are going to be done as separate enhacement tasks in the next sprint, some (repo/packaging,file structure will be on me) another ones by you (I will just guide you with how to further improve it as well as ideas of what we might additionally need there), but that is out of scope for this sprint.
for now the web team is going to use that as a PPA, which is totally fine for now, they might come back to you with some questions/feedback, please provide them with needed support with that. 
I am going to help Robert with deploying the PPA part.

Patch-set: 8