Update patch set 7
Patch Set 7:

> yes i'm not concerned about the regular islation level;
 > 
 > moreover just having another $gard for the next bunch of stmts
 > (even it might be only one one, but you never know how dbix does).
 > 
 > like in NGCP::Panel::Controller::API::TopupCash i undefined $guard
 > demonstratively, to trigger destroy/sope end, then starting over
 > with a new one suing same $guard var.

The point is what we want to see as a transaction, separated statements where update/insert/delete is present or a whole operation ?

By "a whole operation" it means that the whole transaction is rolled back if there is some error in returning a response body in the very end, so the $guard->commit then is the last line of the whole scope.

Not really the case with subcriberregistrations though for the moment, as update_item() also implies an external transaction and the rollback mechanism then should contain a proper way to revert the external transaction as well.

Patch-set: 7
Label: Code-Review=0