Update patch set 8
Patch Set 8:

updated it so tnx_guard_scope only covers and outlines the actual db updates.

Although I would prefer it to cover the whole operation level and roll everything back if there is any error, even in the very last straight forward return headers/body (we do not cover that).
But that should be for another task imho, to refactor the code that way + there should be a solution how to handle rollbacks for external transactions (e.g. kamailio xmlrpc calls) in that case.

Patch-set: 8
Label: Code-Review=0