Update patch set 2
Patch Set 2: Code-Review+2

api tests must fail, because testcases that check for the "lock" pref value in detail were not adopted to see "0" value right after eg. creating the subs, but expect the value to be not present.

let's revert this for now, in favour of a more general appraoch for similar prefs; ie. PATCH "replace" should behave as "add" - at least for prefs

Patch-set: 2
Label: Code-Review=+2