Update patch set 3
Patch Set 3: Code-Review+1

LGTM other than the review of dependencies that you mentioned.

It surprised me that it's this simple, I had the idea that this service would be more complicated.

BTW, does the service code have the changes to notify to systemd that it's ready?  I guess not, if it's still "Type=simple".

Patch-set: 3
Reviewer: Manuel Montecelo <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+1