Update patch set 2
Patch Set 2:

maybe this should be presented to hans-peter as a ppa, while not merging it for now..?

at least to me the current situation seems to overcomplicate things.  considering state variables of controllers is a very last resort, comparing it with passing additional sub params directly in the past. we have special handling here both in old and new structure, while in old structure it was at least visible at first sight...?

Patch-set: 2
Label: Verified=0