Update patch set 2
Patch Set 2:

> ok its for catalyst itself (1. for session store if file is
 > configured and second for upload temp folder), also see eg.
 
(I also suspected of use as upload temp folder :) )


 > sudo rm /tmp/ngcp/panel/session_data
 > sudo rmdir /tmp/ngcp/panel
 > 
 > i'm just concerned about a need to move/reconfigure these locations
 > in panel ..?

 > ok its for catalyst itself (1. for session store if file is
 > configured and second for upload temp folder), also see eg.
 > 
 > sudo rm /tmp/ngcp/panel/session_data
 > sudo rmdir /tmp/ngcp/panel
 > 
 > i'm just concerned about a need to move/reconfigure these locations
 > in panel ..?

This change is about creating the file at boot, rather than at installation time, as it was done before (as part of scripts within the .deb package).

/tmp/ can get cleaned on reboots, and some process (as Alex found out) creates it with the right permissions, so probably creating it at boot is unnecessary.

Probably the original issue was that some other process or script was creating it with the wrong permissions initially, e.g. as root, so the scripts needed to change owner while upgrading.

IMO we can try to just drop the craetion of that file at boot, and if we find any problems, we restore it (either at boot, or when starting the service).

Patch-set: 2
Label: Code-Review=0